Layout the PCB and generate Gerbers.
I am an idiot. This is my first pass at laying out the board.

What I should have done is mark out the edge cuts and then generate a 3d
model to import into Fusion to check that it fits in the case.

I did not do this until after I'd layed out the board :'( and it's about
5mm too long on each side.

So now I need to make the board smaller by a significant amount and redo
the layout.

Think I learned a lot about laying out a PCB though.